# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Finanzas por Voz API
  version: 0.1.0
  description: Contrato preliminar del MVP. Sujeto a validación de Fase 0.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      security: []
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Usuario registrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /auth/login:
    post:
      tags: [Auth]
      security: []
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Sesión iniciada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/refresh:
    post:
      tags: [Auth]
      security: []
      operationId: refreshSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
                  minLength: 20
      responses:
        "200":
          description: Tokens rotados
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    RegisterRequest:
      type: object
      additionalProperties: false
      required: [email, password, full_name, country_code, timezone, default_currency]
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        password:
          type: string
          minLength: 10
          maxLength: 128
        full_name:
          type: string
          minLength: 2
          maxLength: 100
        country_code:
          type: string
          pattern: "^[A-Z]{2}$"
        timezone:
          type: string
          example: America/Bogota
        default_currency:
          type: string
          pattern: "^[A-Z]{3}$"
          example: COP
    LoginRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      additionalProperties: false
      required: [user, tokens]
      properties:
        user:
          $ref: "#/components/schemas/User"
        tokens:
          $ref: "#/components/schemas/TokenPair"
    TokenPair:
      type: object
      additionalProperties: false
      required: [access_token, refresh_token, token_type, expires_in]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          const: bearer
        expires_in:
          type: integer
          minimum: 1
    User:
      type: object
      additionalProperties: false
      required: [id, email, full_name, timezone, default_currency]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        full_name:
          type: string
        timezone:
          type: string
        default_currency:
          type: string
...
    Problem:
      type: object
      additionalProperties: true
      required: [type, title, status, detail, trace_id]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        trace_id:
          type: string
  responses:
    Unauthorized:
      description: Credenciales ausentes o inválidas
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    Conflict:
      description: Conflicto de estado o clave de idempotencia
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    ValidationError:
      description: Entrada inválida
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"